requirements.txt
streamlit
pandas
numpy
yfinance
plotly
scikit-learn
tensorflow-cpu==2.15.0
streamlit-autorefresh